modify action types
