+++ src/actionTypes/account.ts
@@ -1,31 +1,5 @@
-import {
-    INCREMENT,
-    DECREMENT,
-    UPDATE_COUNT,
-    GET_ACCOUNT,
-    UPDATE_ACCOUNT
-} from '~/constant';
+import { GET_ACCOUNT, UPDATE_ACCOUNT } from '~/constant';
 import { Account } from '~/modelTypes';
-
-interface IncrementAction {
-    type: typeof INCREMENT;
-    payload: number;
-}
-
-interface DecrementAction {
-    type: typeof DECREMENT;
-    payload: number;
-}
-
-interface UpdateCountAction {
-    type: typeof UPDATE_COUNT;
-    payload: number;
-}
-
-export type CounterActionTypes =
-    | IncrementAction
-    | DecrementAction
-    | UpdateCountAction;
 
 interface GetAccountAction {
     type: typeof GET_ACCOUNT;
